requirements.txt
python-telegram-bot==13.7
python-dotenv==1.0.1
openai==0.27.0

